LayoutTests/imported/w3c/web-platform-tests/fetch/api/policies/referrer-no-referrer-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'referrer'.


PASS Request's referrer is empty 
